backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic>=2.5.2
pydantic-settings==2.1.0

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
langchain-text-splitters==0.0.1

# Vector database
weaviate-client==3.25.3

# LLM integration
mistralai==0.1.2
openai==1.6.1

# Utilities
aiofiles==23.2.1
httpx==0.25.2
